requirements.txt
flask
openai
torch
transformers
langchain
langchain-openai
langchain-community
python-dotenv
faiss-cpu
beautifulsoup4